+++ packages/client/src/components/common/Icon/Icon.tsx
@@ -7,7 +7,11 @@ const iconPaths = {
   search: 'm21 21l-4.486-4.494M19 10.5a8.5 8.5 0 1 1-17 0a8.5 8.5 0 0 1 17 0Z',
 }
 
-export const User = ({ className = '' }: { className?: string }) => (
+type IconProps = {
+  className?: string
+}
+
+export const User = ({ className = '' }: IconProps) => (
   <svg
     className={`${styles.icon} ${className}`}
     viewBox="0 0 24 24"
@@ -18,7 +22,7 @@ export const User = ({ className = '' }: { className?: string }) => (
   </svg>
 )
 
-export const Mail = ({ className = '' }: { className?: string }) => (
+export const Mail = ({ className = '' }: IconProps) => (
   <svg
     className={`${styles.icon} ${className}`}
     viewBox="0 0 24 24"
@@ -29,7 +33,7 @@ export const Mail = ({ className = '' }: { className?: string }) => (
   </svg>
 )
 
-export const Lock = ({ className = '' }: { className?: string }) => (
+export const Lock = ({ className = '' }: IconProps) => (
   <svg
     className={`${styles.icon} ${className}`}
     viewBox="0 0 24 24"
@@ -40,7 +44,7 @@ export const Lock = ({ className = '' }: { className?: string }) => (
   </svg>
 )
 
-export const Search = ({ className = '' }: { className?: string }) => (
+export const Search = ({ className = '' }: IconProps) => (
   <svg
     className={`${styles.iconLarge} ${className}`}
     xmlns="http://www.w3.org/2000/svg"
@@ -48,12 +52,12 @@ export const Search = ({ className = '' }: { className?: string }) => (
     fill="none"
     stroke="currentColor"
     strokeWidth="2"
-    stroke-linecap="round">
+    strokeLinecap="round">
     <path d={iconPaths.search} />
   </svg>
 )
 
-export const Trophy = ({ className = '' }: { className?: string }) => (
+export const Trophy = ({ className = '' }: IconProps) => (
   <svg
     className={`${styles.icon} ${className}`}
     viewBox="0 0 128 128"
@@ -99,7 +103,7 @@ export const Trophy = ({ className = '' }: { className?: string }) => (
   </svg>
 )
 
-export const InfoTrophy = ({ className = '' }: { className?: string }) => (
+export const InfoTrophy = ({ className = '' }: IconProps) => (
   <svg
     className={`${styles.iconLarge} ${className}`}
     xmlns="http://www.w3.org/2000/svg"
@@ -107,20 +111,20 @@ export const InfoTrophy = ({ className = '' }: { className?: string }) => (
     height="24"
     viewBox="0 0 24 24"
     fill="none"
-    stroke="#cb5300fa"
-    stroke-width="2"
-    stroke-linecap="round"
-    stroke-linejoin="round">
+    stroke="currentColor"
+    strokeWidth="2"
+    strokeLinecap="round"
+    strokeLinejoin="round">
     <path d="M6 9H4.5a2.5 2.5 0 0 1 0-5H6" />
     <path d="M18 9h1.5a2.5 2.5 0 0 0 0-5H18" />
     <path d="M4 22h16" />
     <path d="M10 14.66V17c0 .55-.47.98-.97 1.21C7.85 18.75 7 20.24 7 22" />
     <path d="M14 14.66V17c0 .55.47.98.97 1.21C16.15 18.75 17 20.24 17 22" />
-    <path stroke-linecap="round" d="M18 2H6v7a6 6 0 0 0 12 0V2Z" />
+    <path strokeLinecap="round" d="M18 2H6v7a6 6 0 0 0 12 0V2Z" />
   </svg>
 )
 
-export const InfoZap = ({ className = '' }: { className?: string }) => (
+export const InfoZap = ({ className = '' }: IconProps) => (
   <svg
     className={`${styles.iconLarge} ${className}`}
     xmlns="http://www.w3.org/2000/svg"
@@ -128,15 +132,15 @@ export const InfoZap = ({ className = '' }: { className?: string }) => (
     height="24"
     viewBox="0 0 24 24"
     fill="none"
-    stroke="#cb5300fa"
-    stroke-width="2"
-    stroke-linecap="round"
-    stroke-linejoin="round">
+    stroke="currentColor"
+    strokeWidth="2"
+    strokeLinecap="round"
+    strokeLinejoin="round">
     <polygon points="13 2 3 14 12 14 11 22 21 10 12 10 13 2" />
   </svg>
 )
 
-export const InfoUsers = ({ className = '' }: { className?: string }) => (
+export const InfoUsers = ({ className = '' }: IconProps) => (
   <svg
     className={`${styles.iconLarge} ${className}`}
     xmlns="http://www.w3.org/2000/svg"
@@ -144,10 +148,10 @@ export const InfoUsers = ({ className = '' }: { className?: string }) => (
     height="24"
     viewBox="0 0 24 24"
     fill="none"
-    stroke="#cb5300fa"
-    stroke-width="2"
-    stroke-linecap="round"
-    stroke-linejoin="round">
+    stroke="currentColor"
+    strokeWidth="2"
+    strokeLinecap="round"
+    strokeLinejoin="round">
     <path d="M16 21v-2a4 4 0 0 0-4-4H6a4 4 0 0 0-4 4v2" />
     <circle cx="9" cy="7" r="4" />
     <path d="M22 21v-2a4 4 0 0 0-3-3.87" />
@@ -155,7 +159,7 @@ export const InfoUsers = ({ className = '' }: { className?: string }) => (
   </svg>
 )
 
-export const InfoHeart = ({ className = '' }: { className?: string }) => (
+export const InfoHeart = ({ className = '' }: IconProps) => (
   <svg
     className={`${styles.iconLarge} ${className}`}
     xmlns="http://www.w3.org/2000/svg"
@@ -163,10 +167,10 @@ export const InfoHeart = ({ className = '' }: { className?: string }) => (
     height="24"
     viewBox="0 0 24 24"
     fill="none"
-    stroke="#cb5300fa"
-    stroke-width="2"
-    stroke-linecap="round"
-    stroke-linejoin="round">
+    stroke="currentColor"
+    strokeWidth="2"
+    strokeLinecap="round"
+    strokeLinejoin="round">
     <path d="M19 14c1.49-1.46 3-3.21 3-5.5A5.5 5.5 0 0 0 16.5 3c-1.76 0-3 .5-4.5 2-1.5-1.5-2.74-2-4.5-2A5.5 5.5 0 0 0 2 8.5c0 2.3 1.5 4.05 3 5.5l7 7Z" />
   </svg>
 )
